+++ src/includes/settings/extension.ts
@@ -2,17 +2,43 @@ import {SettingBase} from './base';
 
 
 export class ExtensionSetting extends SettingBase {
-    public filePath: string;
-    public opacity:  number;
+    public filePath:          string;
+	public slideFilePaths:    Array<string>;
+    public opacity:           number;
+	public slideInterval:     number;
+	public slideIntervalUnit: string;
 
     constructor() {
         super('wallpaper-setting');
-        this.filePath = this.get('filePath');
-        this.opacity  = this.get('opacity');
+        this.filePath          = this.get('filePath');
+		this.slideFilePaths    = this.get('slideFilePaths');
+        this.opacity           = this.get('opacity');
+		this.slideInterval     = this.get('slideInterval');
+		this.slideIntervalUnit = this.get('slideIntervalUnit');
     }
+
+	get slideIntervalUnit2Millisecond(): number {
+		let baseTime : number = 1;
+
+		switch(this.slideIntervalUnit) {
+			case "Hour":
+				baseTime *= 60;
+			case "Minute":
+				baseTime *= 60;
+			case "Second":
+				baseTime *= 1000;
+			default:
+				baseTime *= 1;
+		}
+
+		return this.slideInterval * baseTime;
+	}
 
     public uninstall(): void {
         this.remove('filePath');
+		this.remove('slideFilePaths');
         this.remove('opacity');
+		this.remove('slideInterval');
+		this.remove('slideIntervalUnit');
     }
 }