+++ src/includes/utils/base/string.ts
@@ -1,0 +1,8 @@
+export function formatByArray(message: string, ...args: any[]): string {
+	return message.replace(
+		/\{(\d+)\}/g,
+		(blanket, index) => {
+			return args[index];
+		}
+	);
+}