+++ src/includes/wallpaper.ts
@@ -1,5 +1,6 @@
 import * as path          from 'path';
 
+import {formatByArray}    from './utils/base/string';
 import {ExtensionSetting} from './settings/extension';
 import {File}             from './utils/base/file';
 
@@ -26,12 +27,37 @@ export class Wallpaper {
         this.extensionKey    = extensionKey;
     }
 
+	public isInstall(): boolean {
+        let script = new File(
+			this.installPath,
+			{encoding: 'utf-8'}
+		).toString().match(this.getScriptMatch());
+
+		return script ? true : false;
+	}
+
+	public isReady(): boolean {
+		return this.settings.filePath.length > 0 || this.settings.slideFilePaths.length > 0 ? true : false;
+	}
+
     public install(): void {
         let editFile = new File(this.installPath, {encoding: 'utf-8'});
 
         editFile.content = this.clearWallpaperScript(editFile.content) + this.getWallpaperScript(
             this.settings.filePath,
             this.settings.opacity
+        );
+
+        editFile.write({encoding: 'utf-8'});
+    }
+
+    public installAsSlide(): void {
+        let editFile = new File(this.installPath, {encoding: 'utf-8'});
+
+        editFile.content = this.clearWallpaperScript(editFile.content) + this.getSlideScript(
+            this.settings.slideFilePaths,
+            this.settings.opacity,
+			this.settings.slideIntervalUnit2Millisecond
         );
 
         editFile.write({encoding: 'utf-8'});
@@ -54,28 +80,76 @@ export class Wallpaper {
         if(filePath && opacity) {
             let image = new File(filePath);
 
-            result += `
-/*${this.extensionKey}-start*/
-/*${this.extensionKey}.ver.${packageInfo.version}*/
-window.onload=()=>{`;
-            result += `document.body.style.opacity=${this.settings.opacity};`;
-            result += `document.body.style.backgroundSize="cover";`;
-            result += `document.body.style.backgroundPosition="Top Left";`;
-            result += `document.body.style.backgroundRepeat="no-repeat";`;
-            result += `document.body.style.backgroundImage='url("data:image/${image.extension};base64,${image.toBase64()}")';`;
-            result += `}
-/*${this.extensionKey}-end*/`;
+			result = formatByArray(
+				this.getScriptTemplate(opacity),
+				`document.body.style.backgroundImage='url("data:image/${image.extension};base64,${image.toBase64()}")';`
+			);
         }
 
         return result;
     }
 
+	private getSlideScript(
+		filePaths: Array<string>,
+		opacity:   number,
+		interval:  number
+	): string {
+		let result: string = '';
+
+		if(filePaths.length > 0 && opacity && interval) {
+			let temp:  string = `let images=new Array();`;
+			let count: number = 0
+
+			filePaths.forEach(
+				(filePath) => {
+					let image  = new File(filePath);
+					temp      += `images.push('url("data:image/${image.extension};base64,${image.toBase64()}")');`;
+					count     += 1;
+				}
+			)
+
+			temp += `let i=0;`
+			temp += `document.body.style.backgroundImage=images[i];i++;if(i===${count}){i=0;}`
+			temp += `setInterval(`;
+			temp +=     `()=>{`;
+			temp +=         `document.body.style.backgroundImage=images[i];i++;if(i===${count}){i=0;}`;
+		    temp +=     `},`;
+			temp +=     `${interval}`;
+			temp +=     `);`
+
+			result = formatByArray(
+				this.getScriptTemplate(opacity),
+				temp
+			);
+		}
+
+		return result;
+	}
+
+	private getScriptTemplate(opacity: number): string {
+		let result =`
+/*${this.extensionKey}-start*/
+/*${this.extensionKey}.ver.${packageInfo.version}*/
+window.onload=()=>{`;
+					result += `document.body.style.opacity=${opacity};`;
+					result += `document.body.style.backgroundSize="cover";`;
+					result += `document.body.style.backgroundPosition="Top Left";`;
+					result += `document.body.style.backgroundRepeat="no-repeat";`;
+					result += `{0}`;
+					result += `}
+/*${this.extensionKey}-end*/`;
+
+		return result;
+	}
+
     private clearWallpaperScript(content: string): string {
-        let re = new RegExp(
+        return content.replace(this.getScriptMatch(), '').replace(/\s*$/, '');
+    }
+
+	private getScriptMatch(): RegExp {
+        return new RegExp(
             '\\/\\*' + this.extensionKey + '-start\\*\\/[\\s\\S]*?\\/\\*' + this.extensionKey + '-end\\*\\/',
             'g'
         );
-
-        return content.replace(re, '').replace(/\s*$/, '');
-    }
+	}
 }