+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.1.0
+- Change icon
+- Changed from command driven to menu invoked by command
+- Added slide function
+
 ## 0.0.1
 
 - Initial release